Add favorites

+++ app/services/properties_services.py
@@ -1,8 +1,6 @@
 from beanie import Document
-from pydantic import BaseModel
 from app.schemas.properties_schemas import PropertyCreate, PropertyUpdate, PropertyResponse
 from fastapi import HTTPException
-from app.models.nosql.property_model import Property
 from app.crud.properties_crud import (
     create_property_crud,
     get_property_crud,
@@ -10,20 +8,21 @@ from app.crud.properties_crud import (
     show_all_properties_crud,
     update_property_crud,
     delete_property_crud,
+    delete_all_properties_crud
 )
 # create property
 
 
 async def create_property_services(user_id: int, data: PropertyCreate):
 
-    property = await create_property_crud(user_id, data)
+    prop = await create_property_crud(user_id, data)
 
-    return await serialize(property)
+    return await serialize(prop)
 
 # show all user properties
 
 
-async def show_all_user_properties_services(user_id: str):
+async def show_all_user_properties_services(user_id: int):
 
     properties = await show_all_user_properties_crud(user_id)
 
@@ -50,11 +49,14 @@ async def show_all_properties_services():
 
 async def update_property_service(
     property_id: str,
-    user_id: str,
+    user_id: int,
     data: PropertyUpdate
 ):
 
-    property = await get_property_crud(property_id, user_id)
+    prop = await get_property_crud(property_id)
+    if prop.user_id != user_id:
+        raise HTTPException(
+            status_code=401, detail="Unauthorized")
 
     updated_property = await update_property_crud(property, data)
 
@@ -64,47 +66,52 @@ async def update_property_service(
 # delete property
 
 
-async def delete_property_services(property_id: str, user_id: str):
-    property: Property = await get_property_crud(property_id, user_id)
-    is_available = property.status == "available" or "paused"
+async def delete_property_services(property_id: str, user_id: int):
+    prop = await get_property_crud(property_id)
+    if prop.user_id != user_id:
+        raise HTTPException(
+            status_code=401, detail="Unauthorized")
+    is_available = prop.status == "available" or "paused"
     if not is_available:
         raise HTTPException(
             status_code=403, detail="property has active bookings")
     await delete_property_crud(property)
     return {"menssage": "Property deleted successfully"}
 
+# delete all properties
+
+
+async def delete_all_properties_services():
+    await delete_all_properties_crud()
+    return {"menssage": "All deleted successfully"}
+
 
 # pause property
 
 
-async def pause_property(property_id: str, user_id: str):
-    property: Property = await get_property_crud(property_id, user_id)
-    is_reserved = property.status == "reserved"
+async def pause_property(property_id: str, user_id: int):
+    prop = await get_property_crud(property_id)
+    if prop.user_id != user_id:
+        raise HTTPException(
+            status_code=401, detail="Unauthorized")
+    is_reserved = prop.status == "reserved"
     if is_reserved:
         raise HTTPException(
             status_code=403, detail="property has active bookings")
-    is_paused = property.status == "paused"
+    is_paused = prop.status == "paused"
     if is_paused:
         raise HTTPException(
             status_code=409, detail="this property is already paused")
-    await property.set({"status": "paused"})
+    await prop.set({"status": "paused"})
 
     return PropertyResponse(
-        id=str(property.id),
-        user_id=property.user_id,
-        status=property.status
+        id=str(prop.id),
+        user_id=prop.user_id,
+        status=prop.status
     )
-
-
-# Mapear un modelo a un response, ajustando el id
 
 
 async def serialize(object: Document):
-    BaseModel(
-        **object.model_dump(), id=str(object.id)
-    )
-
-# async def serialize(property: Property):
-#     PropertyResponse(
-#         **property.model_dump(), id=str(property.id)
-#     )
+    data = object.model_dump()
+    data["id"] = str(object.id)
+    return data
